Refactor clasp logout test
